Moved Walker impl into walk/mod.rs
